Improve hashing function
